requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.10.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
pyyaml>=6.0
omegaconf>=2.3.0

# Bayesian ML libraries
pyro-ppl>=1.8.0
torchmetrics>=1.0.0

# Visualization and demo
streamlit>=1.25.0
plotly>=5.15.0

# Development tools
black>=23.0.0
ruff>=0.0.280
mypy>=1.5.0
pytest>=7.4.0

# Optional: experiment tracking
wandb>=0.15.0
